# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Todo API",
        "description": "API for managing todos",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/todos": {
            "get": {
                "tags": [
                    "Todo"
                ],
                "summary": "To get all todos",
                "operationId": "57c39ca34a9aced07cae88b3a087f89e",
                "parameters": [
                    {
                        "name": "todo_title",
                        "in": "query",
                        "description": "todo's title",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "is_completed",
                        "in": "query",
                        "description": "indicator of the todo's completeness (true - completed, false - non-complete)",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "sortBy",
                        "in": "query",
                        "description": "based on attribute that return",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "orderBy",
                        "in": "query",
                        "description": "asc / desc",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "401": {
                        "description": "unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "boolean value to determine request success or fail",
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "error message"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "boolean value to determine request success or fail",
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "description": "error message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "boolean value to determine request success or fail",
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/TodoResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Todo"
                ],
                "summary": "To create a new todo",
                "operationId": "b99f2f648b0ae73f79ae59254767d4c1",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TodoCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "boolean value to determine request success or fail",
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "error message"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "boolean value to determine request success or fail",
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "description": "error message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "boolean value to determine request success or fail",
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/TodoResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/todos/{uuid}/status": {
            "patch": {
                "tags": [
                    "Todo"
                ],
                "summary": "To update a todo's status",
                "operationId": "7ba16f58afdbdcdcc1163206c3c01208",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "todo's uuid",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TodoUpdateStatus"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "boolean value to determine request success or fail",
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "error message"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "boolean value to determine request success or fail",
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Todo not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "boolean value to determine request success or fail",
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "description": "error message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "boolean value to determine request success or fail",
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/TodoResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "TodoCreate": {
                "title": "Todo Create",
                "description": "Request body for create todo",
                "required": [
                    "todo_title"
                ],
                "properties": {
                    "todo_title": {
                        "title": "todo_title",
                        "description": "todo's title",
                        "type": "string",
                        "example": "Task 1"
                    }
                },
                "type": "object"
            },
...
            "TodoUpdateStatus": {
                "title": "Todo Update Status",
                "description": "Request body for update todo's status",
                "required": [
                    "is_completed"
                ],
                "properties": {
                    "is_completed": {
                        "title": "is_completed",
                        "description": "indicator of the todo's completeness (true - completed, false - non-complete)",
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
            "TodoResource": {
                "title": "Todo",
                "description": "Todo data",
                "properties": {
                    "uuid": {
                        "title": "uuid",
                        "description": "todo's uuid",
                        "type": "string",
                        "example": "f8928d56-6c89-4c51-881a-9f1ed607f421"
                    },
                    "title": {
                        "title": "title",
                        "description": "todo's title",
                        "type": "string",
                        "example": "Task 1"
                    },
                    "is_completed": {
                        "title": "is_completed",
                        "description": "indicator of the todo's completeness (true - completed, false - non-complete)",
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "title": "created_at",
                        "description": "record creation date time",
                        "type": "string",
                        "example": "2024-01-01 00:00:00"
                    },
                    "updated_at": {
                        "title": "updated_at",
                        "description": "record modification date time",
                        "type": "string",
                        "example": "2024-01-01 00:00:00"
                    }
                },
                "type": "object"
            }
        }
    },
...
}